+++ src/gradient.vala
@@ -1,0 +1,6 @@
+    namespace LiveChart {
+        public struct LinearGradient {
+            public Gdk.RGBA from;
+            public Gdk.RGBA to;
+        }
+    }