+++ tests/grid.vala
@@ -36,5 +36,29 @@ private void register_grid() {
             assert(color.alpha == 1);
         }
     });
+
+    Test.add_func("/LiveChart/Grid/draw#gradient", () => {
+        //Given
+        Cairo.ImageSurface surface = new Cairo.ImageSurface(Cairo.Format.ARGB32, 10, 10);
+        Cairo.Context context = new Cairo.Context(surface);
+        cairo_background(context);
+
+        var grid = new LiveChart.Grid();
+        grid.visible = false;
+        grid.gradient = {from: red(), to: green()};
+
+        //When
+        grid.draw(context, create_config(10, 10));
+ 
+        //Then
+        var pixbuff = Gdk.pixbuf_get_from_surface(surface, 0, 0, 10, 10);
+        if (pixbuff != null) {
+            var at = color_at(pixbuff);
+            assert(at(0, 0).to_string() == "rgb(242,13,0)");
+            assert(at(0, 9).to_string() == "rgb(13,242,0)");            
+        } else {
+            assert_not_reached();
+        }
+    });   
 }